Added sanitize-html and validator dependencies
Added sanitize-html to help validate body, parameters and id for a given endpoint. Helps cover JS, HTML and CSS injections.

Validator is used to check if the URL being passed in is valid, if its not (contains illegal characters or contains some sort of injection) , generates an error and sends it to the user.

sanitize_user_input is the function to sanitize all the content.

Added another middleware for express to check the URL when any request is sent and return an error if the URL is bad. Also sanitizes parameters, query strings and request body on all requests.