Added user login, creation, verification and hashing to backend
Created middleware to authenticate jwt tokens on required endpoints.

Created POST request for login that responds with the user and given token, a POST request to register a user that returns the user and token and a GET request to verify a given user with a verification token sent to their email.

Created user database, hashPassword method that hashes a password with bcrypt (10 rounds) and a method to compare hashed password and given password.